Prod prow config (#2238)
* Copied config from upgrade dev branch

* Moved jobguard to separate jobguard directory.

* moved new version of install-prow.sh to scripts directory

* updated clonerefs image version

* updated docs to match new paths

* Another docs corrections

* Upgrade prow config template.

* generated config from template

* Update readme in old jobguard location
Add exception for building in before-commit.sh

* corrected variable name

* docs update

* test infra revision version update to match components version

* Run dep ensure to update Gopkg.lock

* corrected gcsweb URL to https

* regenerated prow config

* Fixed jobguard package name and import paths

* Update development/jobguard/README.md


* Update development/jobguard/README.md


* Update docs/prow/prow-cluster-update.md


* Update docs/prow/prow-cluster-update.md


* Update docs/prow/prow-cluster-update.md


* Update docs/prow/prow-cluster-update.md


* Update prow/cluster/README.md


* Update development/jobguard/README.md


* reverted files content

* Grammar changes.

* Corrected table formating and content

+++ development/tools/before-commit.sh
@@ -40,6 +40,9 @@ fi
 while IFS= read -r -d '' directory
 do
     cmdName=$(basename "${directory}")
+    if [ -a "${directory}/nobuild.lock" ]; then
+      continue
+    fi
     ${buildEnv} go build -o "${cmdName}" "${directory}"
     buildResult=$?
     rm "${cmdName}"
